The line
${F90_OPTIMISE_FLAGS='$(F77_OPTIMISE_FLAGS)

in the PGI configuration section in lib/make/known-architectures/linux
causes trouble.  It means "use as F90 flags whatever the F77 flags
will be set to".  This doesn't work together with the a recent patch
to configure.in, which says "use the same flags for F77 as for F90".
Overall we now have a circular definition.

As a quick fix, I replace the above line with the value of
F77_OPTIMISE_FLAGS.  Since there is no other code which sets a shell
variable (the left hand side) to a Makefile expression (the right hand
side), it may have been an oversight anyway -- probably the author of
the above line wanted to write ${F77_OPTIMISE_FLAGS} instead anyway.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4184 17b73243-c579-4c4c-a9d2-2d5706c11dac