Apply patch http://www.cactuscode.org/old/pipermail/patches/2005-October/000112.html:
Check if there is more than one assignment to a
parameter in the parfile and will print a warning with line number
information in such a case. It checks both for multiple parameter
assignments of the same or a different value.
Depending on the parameter checking level (as set via the command line
option '-parameter-level {strict|normal|relaxed}') the flesh will also
abort the run after printing the warning in the following two cases:

  * the parameter level is "normal" (which is the default) or "strict"
    and there are multiple assignments of different values

  * the parameter level is "strict" and there are multiple assignments
    of the same value


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4188 17b73243-c579-4c4c-a9d2-2d5706c11dac