Correct handling of newline characters and spaces when writing the configuration bindings.
When reading the configuration settings from the output of the
configuration scripts (that provide capabilities), new lines were lost
in certain circumstances.  This lead to makefile fragments like

PETSC_DIR = /usr/local/apps/petscPETSC_ARCH = linux

which are obviously incorrect (since these should be two lines, not
one), although this does not lead to errors or warnings.

This patch makes sure that all configuration settings have newline
characters appended or blanks prepended as they are concatenated into
the internal database (ConfigScriptParser.pl).  When the database is
read out, the strings can just be printed to the file
(CreateConfigurationBindings.pl).

While doing this, this patch also removes some superfluous blanks and
newline characters from the output of the configuration settings.

This patch also removes a perl idiosyncracy when appending newline
characters to a string.  It changes code like "#endif" . "\n" to
"#endif\n".

Tested on Mac OS and Red Hat Linux.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4199 17b73243-c579-4c4c-a9d2-2d5706c11dac