See Cactus Patches <patches@cactuscode.org>     Subject: CactusDefaultEvolve.c cleanup     Date: Fri, 23 Dec 2005 15:43:50 +0100
The code in CactusDefaultEvolve.c, was complicated by
        1) a global variable, cactus_terminate, which is checked but
           never set except in commented-out code, and
        2) a lot of other commented-out code.

This patch only removes code that is never used, and is intended to make
the remaining code easier to follow and potentially safer.

I checked all arrangements available to me, including Carpet, for
cactus_terminate.  It appears only in
	CactusPUGH/PUGH/src/Evolve.c  (where it is commented out)
	AlphaThorns/Spawner/src/Evolve.c

This change removes

        * live but useless code that references cactus_terminate
        * commented-out TerminationStepper function
        * commented-out code for cactus_terminate_global
        * now-unused TERMINATION_RAISED_BRDCAST
        * numerous commented out CCTK*Timer code

I checked to see that it compiles, and is a patch against the most recent
checkout.  I then ran the standard testsuites with the modified code and
verified that their results were identical to those with the unmodified code.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4253 17b73243-c579-4c4c-a9d2-2d5706c11dac