Application of patch 	http://www.cactuscode.org/old/pipermail/patches/2006-April/000160.html
... a patch for the known-architectures file for Darwin for
the XLF compiler.

Primarily it is to improve the speed, by making the optimization
options to be
	-O3 -qhot -qarch=auto -qtune=auto -qcache=auto
This gets quite good performance, but moves instructions.  The various
"auto's" cause the compiler to generate optimal code for the platform
on which the compiler is running.

Also, I made it supress the congratulatory message ("Yay, I compiled").

Also, I test the compiler version major and minor numbers so that darned
"-no-cpp-precomp" option is included only when the version is less than 3.3.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4313 17b73243-c579-4c4c-a9d2-2d5706c11dac