Application of patch 	GridFuncStuff bug fix and cleanup 	http://www.cactuscode.org/old/pipermail/patches/2006-May/000165.html
It fixes two minor bugs (that is, bugs that weren't bugging anybody right
now).  It also includes an overall cleanup of the code, which begain with
the first bug fix, and resulted in the discovery of the second bug.

The first bug fix was to put in
	use strict;
(According to L. Wall, a Perl file without use strict constitutes a bug.)

This resulted in lots of errors about variables needing their scope to be
specified.  So I did this.

At the end, I found that I couldn't fix a certain occurrance of $group.
There was no sensible way that it was being set.  The code was to produce
a warning concerning creating groups with mixed dimensions.  See
'warn_mixeddim_gfs'.

The Perl code was taking a global value of $group set at build time to be
the last group in a loop.  However, the offending group is determined at
run time in CreateThornGroupInitializers.  So the warning message would
typically be wrong.

Affect code written in
	bindings/Variables/<thorn>.c

Also: Perl 5-ied function calls
      deleted great wads of commented-out code

Testing
-------
Ran testsuites on my laptop with Whisky benchmark thornlist.
No change was detected in the results.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4315 17b73243-c579-4c4c-a9d2-2d5706c11dac