This is the application of the patch reported in Patches and Developsrs Fri 20 Oct 06 "g++ warning: operation on *bla* may be undefined" and PR 2070 http://www.cactuscode.org/cactus_cgi-bin/gnatsweb.pl?cmd=view%20audit-trail&data base=cactus&pr=2070
        Preprocessing arrangements/Carpet/Carpet/src/Timing.cc
	Compiling arrangements/Carpet/Carpet/src/Timing.cc
		configs/chain_gcc_carpet/build/Carpet/Timing.cc:
	In function ‘void Carpet::InitTimingVariables(const cGH*)’:
	configs/chain_gcc_carpet/build/Carpet/Timing.cc:120:
		warning: operation on ‘cctki_use’ may be undefined

The culprit was in generated bindings code meant to get rid of warnings
about unused variables corresponding to Cactus parameters included in
code by the DECLARE_CCTK_PARAMETERS macro.

This replaces the offending code with enums initialised by sizeof calls;
the nature of all the entities involved is much clearer, both to
compilers and to me.

Like the existing solution, this permits the user to proceed to make more
declarations after the DECLARE_CCTK_PARAMETERS macro (simply putting a list
of variable assignments after the declaration would not have that property).


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4381 17b73243-c579-4c4c-a9d2-2d5706c11dac