Make the testsuite mechanism usable for automatic Cactus integration tests (run nightly as non-interactive cron jobs):
1) Via shell environment variables 'CCTK_TESTSUITE_RUN_PROCESSORS' and
'CCTK_TESTSUITE_RUN_COMMAND', the user can override the default number
of processors to use, and the default run command.

2) When running the testsuite scripts, the user is now prompted with a template
for the command to be executed: now an arbitrary command line can be specified
with the number of processors, the executable, and the name of the parameter
file as template parameters.

3) A summary.log file is created in TEST/<configuration>/ subdir containing
the run details of a given configuration and a summary on the number of tests
of tests available/run/passed/failed. This information is the same as printed
to stdout and can be parsed appropriately by an automatic integration test
script.

4) For each testsuite that was run, a <testsuite>.diffs file containing the
differences detected while comparing output files. This information is the
same as printed to stdout and can be parsed appropriately by an automatic
integration test script.

This closes PR Cactus/2072: "enhance testsuite scripts to be used within
automatic Cactus integration tests".


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4391 17b73243-c579-4c4c-a9d2-2d5706c11dac