Correct dependency generation for Fortran files:
Do not set F90_DEPEND, as this variable is not used.  Move definition
of F77_DEPEND to where F_DEPEND is set.

Set F_DEPEND to "$(FPP)", which is evaluated when make runs, not to
the default value of FPP.  This allows influencing F_DEPEND by setting
FPP.

Set AIX_[Ff]90_FLAGS to "-qfree", to be symmetric to AIX_[Ff]_FLAGS
and AIX_[Ff]77_FLAGS, which are set to "-qfixed".


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4411 17b73243-c579-4c4c-a9d2-2d5706c11dac