Fixing Network.c for Cray XT4 Compute Node Linux Compatibility. Network.c uses gethostbyname() which is available on the front-end nodes of the XT4, but not available on Compute Node Linux back-end nodes.  The code will link fine with gethostbyname(), but will resolve to the dynamically shared library version of glibc.  Since the Compute nodes do not understand dynamically linked libraries (only static linking), then it causes a fatal runtime error if gethostbyname() is invoked even though the compiler will not provide any warning that something bad will occur.
Therefore, if you #define CRAY_XT, it will disable the gethostbyname() so as
to prevent a fatal runtime error when GetHostName is called.

Need to find out a way to recognize Cray XT4/Compute-node-Linux systems
automatically.  The front-end of the XT4 looks exactly like a Linux cluster.
(uname returns "Linux").  Perhaps add a test to the Linux cluster known
architectures file that points it to a XT4_CNL known-architetures file.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4434 17b73243-c579-4c4c-a9d2-2d5706c11dac