Add IF clauses to schedule items
Schedule items in Cactus can already have WHILE clauses, which means
that they are executed while a certain condition are true.  This patch
adds IF clauses, which means that they are executed only if a certain
condition is true.  The syntax is equivalent to WHILE clauses.

These if clauses is different from the existing C-syntax if
statements, which are only evaluated at startup and determine how the
schedule is constructed.  These if clauses are evaluated every time
the schedule is traversed, and can skip certain schedule items.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4479 17b73243-c579-4c4c-a9d2-2d5706c11dac