Introduce new configuration variables LIBDIR_PREFIX, RUNDIR_PREFIX, and LIBLINK_PREFIX.  Expand them in makefiles.
LIBDIR_PREFIX and LIBLINK_PREFIX existed before, but were not explicit
configuration variables.  RUNDIR_PREFIX is used to remember the path
to a shared library in an executable.

Keep the default value for LIBDIR_PREFIX at "-L" and LIBLINK_PREFIX at
"-l".  Set the default for RUNDIR_PREFIX to "-Wl,-rpath,".

Use these variables when building the executable.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4599 17b73243-c579-4c4c-a9d2-2d5706c11dac