Patch from Barry Wardell, thanks:
Enables support for compiling HTML versions of
the Cactus documentation. It adds new build targets:

ReferenceManualHTML
MaintainersGuideHTML
UsersGuideHTML
*-ThornDocHTML
*-ArrangementDocHTML
ThornDocHTML
ArrangementDocHTML
ThornGuideHTML

Which are HTML equivalents of the existing doc targets. This makes use
of the small new scripts lib/sbin/ArrangementDocHTML and
lib/sbin/ThornDocHTML.

This patch also introduces a small new script, lib/sbin/ConvertFigures,
which automates the process of converting figures to the correct format for
htlatex.

Finally, it makes some changes to doc/latex/cactus.sty to improve the
rendering of the HTML. It also modifies the Reference Guide to produce a
more consistent result (any cases of the CCTKFunc macro were updated to
use the newer FunctionDescriptionEntry).


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4621 17b73243-c579-4c4c-a9d2-2d5706c11dac