Correct handling multi-line BEGIN DEFINE blocks in configurations scripts
I have discovered a bug in the CreateConfigurationBindings.pl script
that prevents one from declaring more than one definition via the
"BEGIN DEFINE" block in a configuration script. This is because the
perl script receives the entire block as a single string, prefixes it
with "#define" and then writes to the cctki_THORN.h file. I have
modified the script to split the string on newlines and then prefix
each line with "#define", which then provides the expected behavior.

[Patch from Eric L. Seidel]


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4629 17b73243-c579-4c4c-a9d2-2d5706c11dac