Improve optional requirement of capabilities
Thorns can specify that they can make use of a particular capability
(for example HDF5), but that it is not absolutely required. It is
currently not possible to detect a capability in the source code or
make system of a thorn because there is no preprocessor macro and no
make variable defined to indicate that the capability is available.
Additionally, the preprocessor definitions which are explicitly
provided by a capability are written to an include file which is
malformed and hence these definitions cannot be read. This commit:

* Introduces makefile and preprocessor variables HAVE_CAPABILITY_<cap>
  for each provided capability

* Removes the incorrect definition line causing the include file to be
  malformed

* Renames cctki_Capabilities.h to cctk_Capabilities.h and causes
  cctk_Capabilities.h to be included from cctk.h by all thorns

* Excludes cctk_Capabilities.h from dependency checking (dependencies
  of the files included from this one will be sufficient)

[Patch from Ian Hinder]


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4659 17b73243-c579-4c4c-a9d2-2d5706c11dac