Improve auto-detection of "restrict" keyword
Update and improve the auto-detection of the "restrict" keyword; add
more test cases to detect more possible failures. Disable testing the
array syntax

void foo (double A[restrict]);

where the "restrict" keyword is given within the square brackets;
instead, one can use the pointer syntax

void foo (double *restrict const A);

which is tested.

Add support for the gcc built-in function "builtin_expect", which
tells the compiler the value that an expression is most likely to
have.

Re-enable the (currently commented out) definitions for attribute(hot)
and attribute(cold).


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4702 17b73243-c579-4c4c-a9d2-2d5706c11dac