Depreciate config files in test/ directories.
If for some reason there exists both a test.ccl and a file 'config' in a
thorn's test directory, Cactus uses the config file while telling the user that
it's using the test.ccl. Config is so outdated this shouldn't affect anybody.
This patch depreciates config files. Support for test.ccl was added in 2002 and
I coulnd't find a single thorn which still uses config files.



git-svn-id: https://svn.cactuscode.org/flesh/trunk@4733 17b73243-c579-4c4c-a9d2-2d5706c11dac