doc: Remove revision and date from front page in manuals
This was giving the revision and date of the UsersGuide.tex file etc,
and not the date of the last edit to the documentation.  Hence it was
giving the impression that our documentation came from 2004.  Better
to have no date.

This technique also worked only in SVN, and stopped people from using
git-svn.

git-svn-id: https://svn.cactuscode.org/flesh/trunk@4758 17b73243-c579-4c4c-a9d2-2d5706c11dac