Allow zero timelevels in STORAGE timelevels
This allows to simply set timelevels=0 to turn off storage without
having to put an "if (do_something timelevels > 0)" into schedule.ccl
files. It is also the only way to turn off storage inside of a GROUP
of SCHEDULE statement based on a condition (other than scheduling the
item twice, once with STORAGE, once without).


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4774 17b73243-c579-4c4c-a9d2-2d5706c11dac