have CCTK_RegexMatch return a distinct error code if patterns is invalid
This patch changes the return value in the "does not compile" case to -1 and
updates all source files that I could find that use it.

Note that this patch changes behaviour of a routine. It used to return 0 for
non-compiling patterns so thorns that test for C-like true would interpret
invalid patterns as does-not-match, but will interpret the -1 return value as
does-match.



git-svn-id: https://svn.cactuscode.org/flesh/trunk@4831 17b73243-c579-4c4c-a9d2-2d5706c11dac