add parameters to reverse un-enforced Cactus schedule ordering
This patch adds two new parameters to the flesh:

 *  schedule_sort_mode - which affects the order of schedule items before
    enforcing BEFORE/AFTER
 *  schedule_sort_warnings - which outputs warnings if a schedule item refers
    to a non-existing item in its BEFORE/AFTER modifiers

Neither one is intended to be used in production runs but they are useful for
debugging a schedule.
schedule_sort_warnings is intended to catch typos in dependency names and when
one tries to order with respect to items hidden within a group. It will find a
number of false positives for items that are only scheduled based on parameter
settings. Eg. MoL's RHS NaN checker.
schedule_sort_mode can be used to ensure that the schedule order does not
depend on the (semi-random) order that Cactus generates for non-enforced
ordering.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@4845 17b73243-c579-4c4c-a9d2-2d5706c11dac