ParseFile.c: Change semantics of call to fread so that error check makes sense
fread takes an "item size" and a "number of items" and returns the number of items successfully read.  Previously, we asked for 1 item of the size of the file, to read the entire file.  The error check which was added in r4915 assumed that the returned value of fread was the number of bytes read, which is not correct since only one item is returned, and causes Cactus to fail with a fatal error when any parameter file is loaded.  The present commit changes the fread call to ask for file_size items of size 1, for which the error-checking code makes sense, as there are now file_size items returned.

git-svn-id: https://svn.cactuscode.org/flesh/trunk@4917 17b73243-c579-4c4c-a9d2-2d5706c11dac