Currently, known_architectures/linux only looks at the first line of '$F77 --version' to determine the version of a gnu-type fortran compiler. However, on bluewaters this returns
{{{
/opt/cray/xt-asyncpe/5.16/bin/ftn: INFO: Compiling with CRAYPE_COMPILE_TARGET=native.
GNU Fortran (GCC) 4.7.2 20120920 (Cray Inc.)
Copyright (C) 2012 Free Software Foundation, Inc.
}}}

The information the script is actually interested in is on the second line, not
the first. Thus, I propose the attached patch (grepping for "GNU" before
choosing the first line). I didn't find a problem with this patch on other
machines, but there might be in case $F77 --version does not contain GNU in the
line with the version.



git-svn-id: https://svn.cactuscode.org/flesh/trunk@4958 17b73243-c579-4c4c-a9d2-2d5706c11dac