check that providing function name is different from aliased function name
in interface.ccl. Before f an aliased function is declared in interface.ccl with

PROVIDES FUNCTION fun WITH fun LANGUAGE C

 Cactus went into an infinite loop when calling that function. It did so
because it creates a function 'fun' (the first) itself, which then calls (fun)
the second.

 Instead, Cactus now produces an error for above declaration. The two 'fun'
must be different for this to work, and Cactus should catch this.

Reported by Frank Loeffler in ticket 1376.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@5024 17b73243-c579-4c4c-a9d2-2d5706c11dac