Currently Cactus sets up flags like CPPFLAGS or CFLAGS by adding e.g. CPP_OPENMP_FLAGS. However, later it overwrites these again by their original value in sbin/ProcessConfiguration.pl (search for FIXME).
This patch implements what the 'FIXME' suggests - accepting the drawbacks that
are mentioned there: that configuration settings not originating from a thorn
might not be forwarded from e.g., a .cactus/config file. MPI was one of these,
but this is now handled differently anyway. With this patch, we would need to
be aware of these and might need to add them to @allowed_opts in the future.

Without the patch however, compilation might fail for perfectly valid setups.
One of these is when using openmp, setting all the corresponding
*_OPENMP_FLAGS, but not setting CPPFLAGS (only CPP_OPENMP_FLAGS). In this case
ProcessConfiguration.pl will set CFLAGS to the version in the config file
(*without* the -openmp), but it will leave CPPFLAGS to the version *with*
-openmp. This later leads to a linker error in external libraries, since
compilation there uses CPPFLAGS (with openmp), but the linker doesn't (It
correctly uses CFLAGS, but this doesn't have openmp flags here). 



git-svn-id: https://svn.cactuscode.org/flesh/trunk@5034 17b73243-c579-4c4c-a9d2-2d5706c11dac