recover parameters early
CCTK_RECOVER_PARAMETER is special anyway and scheduled very early (even before
the thorn's startup routines are called). This patch to the flesh moves
parameter recovery to just after the parameters from the parfile are processed.
This ensures that anything afterwards sees the parameters from the checkpoint
(for example the cache setup routine which looks at some Cactus parameters and
importantly CCTKi_BindingsVariablesInitialise.


git-svn-id: https://svn.cactuscode.org/flesh/trunk@5114 17b73243-c579-4c4c-a9d2-2d5706c11dac