WarnLevel.c: Pass EXIT_FAILURE to CCTK_Abort
This is notionally the exit code of the process, and it makes more sense
to use EXIT_FAILURE here than EXIT_SUCCESS (0).  Note that CCTK_Abort
ignores this if asserts are enabled, as it just triggers an assert, so
this change should have no effect on the operation of the system in most
cases.

Fixes https://trac.einsteintoolkit.org/ticket/1689.