Output time stamp in verbose build log
Timestamp includes sub-second granularity when available, and uses Perl
to do so in a more portable way than 'date' likely could.

At the same time, don't let 'make' output the commands necessary to
generate this debug output. This is not useful, except when debugging
the debug output, which likely a user will never do.