Fixes issue https://bitbucket.org/einsteintoolkit/tickets/issues/2369/presync-circumvents-scoping-of-cactus-grid
As implemented, the header generated for DECLARE_CCTK_ARGUMENTS_func does not check to see whether the thorn has access to a GF that it wants to read or write. In Fortran, this will lead to a compile time error, but in C/C++, the header will simply grant access to the variable.

Note that CST_Error is reported at level 1. This should be upgraded to level 0 before the next release.