Cactus: lookup_var should return a shared_ptr<const Value>, not a shared_ptr<Value>.
Currently, lookup_var returns a reference to the value of a var. When a unop is called, this has the effect of negating the value in the table. Thus:

$zip = 1
foo::xmin = -$zip
foo::xmax =  $zip

Results in both xmin and xmax being set to the same value (-1). The fix is to have lookup_var return a copy rather than a reference.